unit tests + add db context interface

+++ Totalizator/Services/BetService.cs
@@ -5,14 +5,15 @@ using System.Linq;
 using System.Web;
 using Totalizator.Models;
 using Totalizator.Models.DbModel;
+using Totalizator.Util.Exeptions;
 
 namespace Totalizator.Services
 {
     public class BetService : IBetRepository
     {
-        totalizatorEntities db;
+        ITotalizatorContext db;
 
-        public BetService(totalizatorEntities dbContext)
+        public BetService(ITotalizatorContext dbContext)
         {
             db = dbContext;
         }
@@ -28,6 +29,7 @@ namespace Totalizator.Services
 
         public void DeleteBet(int betId)
         {
+            if (betId <= 0) throw new WrongIdExeption("Id of Bet can't be 0 of lower");
             db.Bets.Remove(db.Bets.Where(p => p.Id == betId).FirstOrDefault());
             db.SaveChanges();
         }
@@ -39,13 +41,13 @@ namespace Totalizator.Services
 
         public IEnumerable<Bet> ListBet()
         {
-            db.Configuration.LazyLoadingEnabled = false;
             return db.Bets;
         }
 
         public IQueryable<Bet> ListBet(int pageNumber, int userId)
         {
-            db.Configuration.LazyLoadingEnabled = false;
+            if (userId <= 0) throw new WrongIdExeption("User Id can't be 0 or lower");
+            if (pageNumber <= 0) throw new WrongPageNumberExeption("Page number can't be 0 or lower");
             int pageSize = 3;
             return db.Bets.Include("Event").Include("Team").Where(x => x.UserId == userId).OrderByDescending(p => p.Event.Date).Skip((pageNumber - 1) * pageSize).Take(pageSize);
         }
